+++ src/buf/impl_grid.rs
@@ -1,6 +1,8 @@
+use ixy::index::Layout;
+
 use crate::{
     buf::GridBuf,
-    core::{Layout, Pos},
+    core::Pos,
     grid::{BoundedGrid, GridBase, GridReadUnchecked, GridWriteUnchecked},
 };
 
@@ -30,18 +32,18 @@ where
     L: Layout,
 {
     unsafe fn get_unchecked(&self, pos: Pos) -> &T {
-        let index = L::to_1d(pos, self.width).index;
+        let index = L::to_1d(pos, self.width);
         unsafe { self.buffer.as_ref().get_unchecked(index) }
     }
 
-    unsafe fn rect_iter_unchecked(
+    unsafe fn iter_rect_unchecked(
         &self,
         bounds: crate::core::Rect,
     ) -> impl Iterator<Item = &Self::Element> {
         let slice = self.buffer.as_ref();
         let width = self.width;
         (bounds.top()..bounds.bottom()).flat_map(move |y| {
-            let row_start = L::to_1d(Pos::new(bounds.left(), y), width).index;
+            let row_start = L::to_1d(Pos::new(bounds.left(), y), width);
             slice[row_start..row_start + bounds.width()].iter()
         })
     }
@@ -53,7 +55,7 @@ where
     L: Layout,
 {
     unsafe fn set_unchecked(&mut self, pos: Pos, value: T) {
-        let index = L::to_1d(pos, self.width).index;
+        let index = L::to_1d(pos, self.width);
         unsafe { *self.buffer.as_mut().get_unchecked_mut(index) = value }
     }
 
@@ -66,7 +68,7 @@ where
         let width = self.width;
         let mut iter = iter.into_iter();
         for y in bounds.top()..bounds.bottom() {
-            let x_xtart = L::to_1d(Pos::new(bounds.left(), y), width).index;
+            let x_xtart = L::to_1d(Pos::new(bounds.left(), y), width);
             let x_end = x_xtart + bounds.width();
             slice[x_xtart..x_end]
                 .iter_mut()
@@ -82,7 +84,7 @@ where
         let slice = self.buffer.as_mut();
         let width = self.width;
         for y in bounds.top()..bounds.bottom() {
-            let x_start = L::to_1d(Pos::new(bounds.left(), y), width).index;
+            let x_start = L::to_1d(Pos::new(bounds.left(), y), width);
             let x_end = x_start + bounds.width();
             slice[x_start..x_end].fill(value);
         }