+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2025-07-19
+
+### Changed
+
+- Renamed various methods like `rect_iter` to `iter_rect` for consistency
+- Renamed `fill_rect_iter` to `fill_rect_from`
+
+### Removed
+
+- No longer exporting `core::Layout`; use `ixy::index::Layout` as needed.
+
 ## [0.3.0] - 2025-07-19
 
 ### Added